fix: move script api to enterprise

+++ packages/nocodb/src/ee/services/v3/scripts-v3.service.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@nestjs/common';
+import { PlanFeatureTypes } from 'nocodb-sdk';
 import type { NcContext, NcRequest } from 'nocodb-sdk';
 import type {
   ScriptV3GetResponseType,
@@ -9,6 +10,8 @@ import type { Script } from '~/models';
 import { ScriptsService } from '~/services/scripts.service';
 import { builderGenerator } from '~/utils/api-v3-data-transformation.builder';
 import { validatePayload } from '~/helpers';
+import { getFeature } from '~/helpers/paymentHelpers';
+import { NcError } from '~/helpers/catchError';
 
 const scriptBuilder = builderGenerator<Script, ScriptV3GetResponseType>({
   allowed: [
@@ -42,7 +45,26 @@ const scriptListItemBuilder = builderGenerator<
 export class ScriptsV3Service {
   constructor(private readonly scriptsService: ScriptsService) {}
 
+  /**
+   * Validates if the user has access to the Scripts API.
+   * This method checks if the feature is enabled for the workspace.
+   * If not, it throws an error indicating that the feature is only available on Enterprise plans.
+   */
+  private async validateFeatureAccess(context: NcContext) {
+    if (
+      !(await getFeature(
+        PlanFeatureTypes.FEATURE_API_SCRIPT_MANAGEMENT,
+        context.workspace_id,
+      ))
+    ) {
+      NcError.get(context).forbidden(
+        'Accessing Scripts API is only available on Enterprise plans. Please upgrade your workspace plan to enable this feature. Your current plan is not sufficient.',
+      );
+    }
+  }
+
   async scriptList(context: NcContext): Promise<ScriptV3ListResponseType> {
+    await this.validateFeatureAccess(context);
     const scripts = await this.scriptsService.listScripts(context);
 
     return {
@@ -54,6 +76,7 @@ export class ScriptsV3Service {
     context: NcContext,
     id: string,
   ): Promise<ScriptV3GetResponseType> {
+    await this.validateFeatureAccess(context);
     const script = await this.scriptsService.getScript(context, id);
     return scriptBuilder().build(script);
   }
@@ -63,6 +86,7 @@ export class ScriptsV3Service {
     body: ScriptV3RequestType,
     req: NcRequest,
   ): Promise<ScriptV3GetResponseType> {
+    await this.validateFeatureAccess(context);
     validatePayload(
       'swagger-v3.json#/components/schemas/ScriptCreateReq',
       body,
@@ -79,6 +103,7 @@ export class ScriptsV3Service {
     body: ScriptV3RequestType,
     req: NcRequest,
   ): Promise<ScriptV3GetResponseType> {
+    await this.validateFeatureAccess(context);
     validatePayload(
       'swagger-v3.json#/components/schemas/ScriptUpdateReq',
       body,
@@ -99,6 +124,7 @@ export class ScriptsV3Service {
     id: string,
     req: NcRequest,
   ): Promise<boolean> {
+    await this.validateFeatureAccess(context);
     return await this.scriptsService.deleteScript(context, id, req);
   }
 }
